:recycle: add service layer

+++ src/main/java/com/trifork/nedgia/dto/PredictedConsumptionResponse.java
@@ -16,6 +16,6 @@ import lombok.Value;
 @JsonNaming(SnakeCaseStrategy.class)
 public class PredictedConsumptionResponse {
 
-    Long predicted;
+    Double predicted;
 
 }
